finished stable version of camera

+++ engine/object.cpp
@@ -6,7 +6,7 @@ Object::Object() {
 Object::~Object() {}
 void Object::setImage(string file, SDL_Renderer* ren) {img.loadImage(file, ren);}
 SDL_Texture* Object::getImage() {return img.getImage();}
-void Object::setSource(int x, int y, int w, int h) {
+void Object::setSource(double x, double y, int w, int h) {
   setSX(x);
   setSY(y);
   setSW(w);
@@ -16,11 +16,11 @@ void Object::setDest(int w, int h) {
   setDW(w);
   setDH(h);
 }
-void Object::setDest(int w, int h, int x, int y) {
+void Object::setDest(int w, int h, double x, double y) {
   setDest(w, h);
   setDestCoord(x, y);
 }
-void Object::setDestCoord(int x, int y) {
+void Object::setDestCoord(double x, double y) {
    setDX(x);
    setDY(y);
 }
@@ -30,29 +30,29 @@ SDL_Rect Object::getSource() {
 SDL_Rect Object::getDest() {
   return dest;
 }
-void Object::setSX(int x) {rect.x = x;}
-void Object::setSY(int y) {rect.y = y;}
+void Object::setSX(double x) {rect.x = x;}
+void Object::setSY(double y) {rect.y = y;}
 void Object::setSW(int w) {rect.w = w;}
 void Object::setSH(int h) {rect.h = h;}
-void Object::setDX(int x) {dest.x = x;}
-void Object::setDY(int y) {dest.y = y;}
+void Object::setDX(double x) {dest.x = x;}
+void Object::setDY(double y) {dest.y = y;}
 void Object::setDW(int w) {dest.w = w;}
 void Object::setDH(int h) {dest.h = h;}
-int Object::getSX() {return rect.x;}
-int Object::getSY() {return rect.y;}
-int Object::getSW() {return rect.w;}
-int Object::getSH() {return rect.h;}
-int Object::getDX() {return dest.x;}
-int Object::getDY() {return dest.y;}
-int Object::getDW() {return dest.w;}
-int Object::getDH() {return dest.h;}
-void Object::setAng(int a) {
+double Object::getSX() {return rect.x;}
+double Object::getSY() {return rect.y;}
+double Object::getSW() {return rect.w;}
+double Object::getSH() {return rect.h;}
+double Object::getDX() {return dest.x;}
+double Object::getDY() {return dest.y;}
+double Object::getDW() {return dest.w;}
+double Object::getDH() {return dest.h;}
+void Object::setAng(double a) {
   angle = a;
 }
-int Object::getAng() {
+double Object::getAng() {
   return angle;
 }
-void Object::move(int mx, int my) {
+void Object::move(double mx, double my) {
   setDX(getDX()+mx);
   setDY(getDY()-my);
 }
